refactor: The code has been made more readable. Files are separated in a more organized way.
