+++ Handlers/Forms.go
@@ -1,0 +1,9 @@
+package Handlers
+
+type DiscordAccessTokenResult struct {
+	AccessToken string `json:"access_token"`
+}
+
+type DiscordUserResult struct {
+	Id string `json:"id"`
+}
